multiplying program work, not tested the rest

+++ CPU6502.h
@@ -191,7 +191,7 @@
 
         uint8_t fetched = 0x00;
         uint16_t addr_abs = 0x0000;
-        uint8_t addr_rel = 0x00;
+        uint16_t addr_rel = 0x00;
         uint8_t opcode = 0x00;
         uint8_t cycles = 0;
         uint16_t temp = 0x0000;
